requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
coverage==7.2.7
dj-database-url==2.1.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-authentication==0.1.1
djangorestframework-simplejwt==5.2.2
freezegun==1.2.2
gunicorn==21.2.0
idna==3.4
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==6.3.0
typing_extensions==4.7.1
urllib3==2.0.4
whitenoise==6.5.0
django-cors-headers==4.2.0